fix broken graphql after /reload (#73)
* add test to catch broken reload

* fix broken graphql after /reload

After upgrading the apollo-server-express version, any queries that
include interfaces will fail after reloading the server.

In order to reproduce we can just make a query like this one:

```
{
  permissions_v1 (path:"<path>") {
    service
    ... on PermissionGithubOrgTeam_v1 {
      org
    }
  }
}
```

Before reloading the object will contain `org`, but once we reload the
`org` parameter will be missing from the response.

This is because with the new apollo-server-express version besides
resetting the schema on a hot reload, we also need to make sure the
`schemaDerivedData` is also generated.

The solution to this issue was found here:
https://github.com/apollographql/apollo-server/issues/1275#issuecomment-532183702

Note the `@ts-ignore` line, which is needed because we are accessing a
private method `generateSchemaDerivedData`. Unless we have this line
typescript will raise an error when compiling.

This PR also adds a test that will catch this if there is a regression.