Adds an extension that returns all traversed schemas
The GraphQL response payload is documented here:
https://spec.graphql.org/June2018/#sec-Response-Format

In our integrations we were relying on only `data` and `errors`, but now
we require a complete list of schemas that each query has traverse, to
identify schema dependencies per query.

The spec supports creating extensions (plugins) to add additional custom
data to the response payload.

With this functionality the response payload will be:

```json
{
  "data": ...,
  "errors": ...,
  "extensions": {
    "schemas": ["/schema1.yaml",...]
  }
}
```