provide a `diff/$old_bundle/$new_bundle` endpoint on qontract-server that provides detailed diffs between two bundles currently present in the qontract-server
this endpoint will be used during PR checks to find involved schemas for an app-interface MR. the diff endpoint needs to expose diffs for resource files as well along with their relation to the referencing data file. this way, we don't have to guess the involved schemas anymore during PR checks but have an explicit and correct way to find them all.

The endpoint returns a list of changes with the following fields

* datafilepath - the path of the data file that changed
* datafileschema - the schema of the datafile that changed
* resourcepath - the path of the resource that changed in the context of the datafile
* action:
** `E` if a field changed between the two bundles
** `A` if a field was added in the new bundle
** `D` if a field was deleted in the new bundle
* jsonpath: the jsonpath pointing to the field or subtree that changed in the datafile
* new - holds the data object from the new bundle - empty if a data file is deleted
* old - holds the data object from the old bundle - empty if a new data file is introduced

resource file changes are reported in the context of their referencing data file.

```yaml
[
    {
      "datafilepath": "/services/api-designer/cicd/ci-int/secrets.yaml",
      "datafileschema": "/dependencies/jenkins-config-1.yml",
      "action": "E",
      "jsonpath": "config_path",
      "resourcepath": "/jenkins/api-designer/secrets.yaml",
      "old": { ... },
      "new": { ... }
    }
    {
      "datafilepath": "/integrations/user-validator.yml",
      "datafileschema": "/app-sre/integration-1.yml",
      "action": "E",
      "jsonpath": "upstream",
      "old": { ... },
      "new": { ... }
   }
]
```

depends loosely on https://github.com/app-sre/qontract-schemas/pull/160 and https://github.com/app-sre/qontract-validator/pull/37. if backrefs are not present in the bundle, the diff endpoint only exposes differences in data

part of https://issues.redhat.com/browse/APPSRE-5629

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>