INIT_BUNDLES from s3 (#146)
renamed INIT_DISK_BUNDLES to INIT_BUNDLES and allow pre-populating a qontract-server with bundles from fs or s3

`INIT_BUNDLES` is a CSV of bundle references of the following formats

* fs:\\path\to\bundle
* s3:\\bundle-key

The s3 flavour relies on the `AWS_*` env variables to specify the bucket and configure authentication. the specified `bundle-key` is used as `AWS_S3_KEY`.

The bundles listed in `INIT_BUNDLE` can come from different sources and are added to the `qontract-server` in the order they are specified. This means that the bundle listed last is also the one returned by the `/sha256` endpoint.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>