synthetic subAttr can references nested fields
the `subAttr` field of a synthetic declaration is used as a foreign key to define the backreference relation to another schema.
this change enables the `subAttr` field to declare this relation not onlyto top level field of the referenced schema but also nested fields.

a `.` (dot)  is used as a separator for nesting levels.

Example

```
- name: StatusPage_v1
  fields:
  - name: components
    type: StatusPageComponent_v1
    isList: true
    synthetic:
      schema: /dependencies/status-page-component-1.yml
      subAttr: some.path.to.a.page
```

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>