simplify diff exposure between two bundles
javascript diffing capabilities are limited and some corner cases can't be covered to detect proper structural diffs, e.g. adding or removing array elements in places other than the end of the array yields a lot of fake changes.

therefore the diff endpoint output has been simplified. it does not report fine grained changes for datafiles and resources anymore but instead exposes only the changes files with some metadata. it is up to the user of the diff endpoint to look for specific changes within the old and new state of a datafile or resource.

this is a breaking change for the endpoint because the change path details have been removed. but since this endpoint is not in use yet, changing it is legit.

in python the deep-diff library is capable to detect all those changes in a stable way, exposing

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>